requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==1.5.0
blinker==1.4
blis==0.4.1
cachetools==4.0.0
catalogue==0.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cmake==3.17.3
cycler==0.10.0
cymem==2.0.3
decorator==4.4.1
defusedxml==0.6.0
dlib==19.19.0
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gast==0.2.2
gevent==1.4.0
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.9999999
idna==2.8
importlib-metadata==1.3.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lab==5.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.0
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.13.0
spacy==2.2.3
SQLAlchemy==1.3.12
srsly==0.2.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.7.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
tornado==6.0.3
tqdm==4.41.0
traitlets==4.3.3
urllib3==1.25.7
wasabi==0.5.0
wcwidth==0.1.7
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
xgboost==0.90
zipp==0.6.0